[preview] run the telemetry job right after gitpod is ready
This PR updates the `entry-point.sh` to run a parallel step
that waits for the gitpod pods to be ready, and manually executes
the cronjob once its necessary.

This is necessary as our telemetry `cronjobs` are only ran
daily in the midnight, and we can't expect this schedule to work
for local previews as they are short lived.

example data that we get:

```
client.Track(&analytics.Track{
  UserId: "334c379a-afee-4882-ba44-fbd7dfc0a8b0",
  Event: "Installation telemetry",
  Properties: map[string]interface{}{
    "customerID": "",
    "platform": "local-preview",
    "totalInstances": 0,
    "totalUsers": 0,
    "totalWorkspaces": 0,
    "version": "tar-preview-tel-init.4",
  },
})
```

Signed-off-by: Tarun Pothulapati <tarun@gitpod.io>